+++ docs/docs-new/pages/guides/_meta.js
@@ -1,0 +1,4 @@
+module.exports = {
+  "examples": "Examples",
+  "recipes": "Recipes"
+}