+++ docs/docs-new/pages/guides/recipes/access-control/_meta.js
@@ -1,0 +1,7 @@
+module.exports = {
+  "enforcing-mandatory-filters": "Enforcing mandatory filters",
+  "column-based-access": "Enforcing column-based access",
+  "role-based-access": "Enforcing role-based access",
+  "controlling-access-to-cubes-and-views": "Controlling access to cubes and views",
+  "using-different-schemas-for-tenants": "Using different data models for tenants"
+}