+++ docs/docs-new/pages/product/configuration/_meta.js
@@ -1,5 +1,4 @@
 module.exports = {
-  "overview": "Overview",
   "data-sources": "Connecting to data sources",
   "visualization-tools": "Connecting to visualization tools",
   "vpc": "Connecting with a VPC",