+++ packages/cubejs-backend-native/src/python/template/mj_value.rs
@@ -1,7 +1,74 @@
 use crate::python::cross::{CLRepr, CLReprKind, CLReprObject};
+use crate::python::runtime::py_runtime;
+use crate::tokio_runtime;
 use minijinja as mj;
-use minijinja::value::{Object, ObjectKind, SeqObject, StructObject, Value};
+use minijinja::value::{Object, ObjectKind, SeqObject, StructObject, Value, ValueKind};
+use pyo3::types::PyFunction;
+use pyo3::Py;
+use std::convert::TryInto;
 use std::sync::Arc;
+
+struct JinjaPythonFunction {
+    pub(crate) inner: Py<PyFunction>,
+}
+
+impl std::fmt::Debug for JinjaPythonFunction {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        std::fmt::Debug::fmt(&self.inner, f)
+    }
+}
+
+impl std::fmt::Display for JinjaPythonFunction {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        std::fmt::Display::fmt(&self.inner, f)
+    }
+}
+
+impl Object for JinjaPythonFunction {
+    fn kind(&self) -> ObjectKind<'_> {
+        ObjectKind::Plain
+    }
+
+    fn call_method(
+        &self,
+        _state: &mj::State,
+        _name: &str,
+        _args: &[Value],
+    ) -> Result<Value, mj::Error> {
+        Err(mj::Error::new(
+            minijinja::ErrorKind::InvalidOperation,
+            "Unable to call method on Py<Function>",
+        ))
+    }
+
+    fn call(&self, _state: &mj::State, args: &[Value]) -> Result<Value, mj::Error> {
+        let mut arguments = Vec::with_capacity(args.len());
+
+        for arg in args {
+            arguments.push(from_minijinja_value(arg)?);
+        }
+
+        let py_runtime = py_runtime()
+            .map_err(|err| mj::Error::new(mj::ErrorKind::EvalBlock, format!("Error: {}", err)))?;
+        let call_future = py_runtime.call_async(self.inner.clone(), arguments);
+
+        let tokio = tokio_runtime()
+            .map_err(|err| mj::Error::new(mj::ErrorKind::EvalBlock, format!("Error: {}", err)))?;
+        match tokio.block_on(call_future) {
+            Ok(r) => Ok(to_minijinja_value(r)),
+            Err(err) => Err(mj::Error::new(
+                mj::ErrorKind::EvalBlock,
+                format!("Call error: {}", err),
+            )),
+        }
+    }
+}
+
+impl StructObject for JinjaPythonFunction {
+    fn get_field(&self, _name: &str) -> Option<Value> {
+        None
+    }
+}
 
 struct JinjaDynamicObject {
     pub(crate) inner: CLReprObject,
@@ -24,11 +91,8 @@ impl StructObject for JinjaDynamicObject {
         self.inner.get(name).map(|v| to_minijinja_value(v.clone()))
     }
 
-    fn fields(&self) -> Vec<Arc<String>> {
-        self.inner
-            .iter()
-            .map(|(k, _)| Arc::new(k.clone()))
-            .collect()
+    fn fields(&self) -> Vec<Arc<str>> {
+        self.inner.keys().map(|x| x.to_string().into()).collect()
     }
 
     fn field_count(&self) -> usize {
@@ -113,6 +177,37 @@ impl Object for JinjaSequenceObject {
     }
 }
 
+pub fn from_minijinja_value(from: &mj::value::Value) -> Result<CLRepr, mj::Error> {
+    match from.kind() {
+        ValueKind::Undefined | ValueKind::None => Ok(CLRepr::Null),
+        ValueKind::Bool => Ok(CLRepr::Bool(from.is_true())),
+        ValueKind::Number => {
+            if let Ok(rv) = TryInto::<i64>::try_into(from.clone()) {
+                Ok(CLRepr::Int(rv))
+            } else if let Ok(rv) = TryInto::<f64>::try_into(from.clone()) {
+                Ok(CLRepr::Float(rv))
+            } else {
+                Err(mj::Error::new(
+                    mj::ErrorKind::InvalidOperation,
+                    format!("Converting from {:?} to python is not supported", from),
+                ))
+            }
+        }
+        ValueKind::String => {
+            if from.is_safe() {
+                // TODO: Danger?
+                Ok(CLRepr::String(from.as_str().unwrap().to_string()))
+            } else {
+                Ok(CLRepr::String(from.as_str().unwrap().to_string()))
+            }
+        }
+        other => Err(mj::Error::new(
+            mj::ErrorKind::InvalidOperation,
+            format!("Converting from {:?} to python is not supported", other),
+        )),
+    }
+}
+
 pub fn to_minijinja_value(from: CLRepr) -> Value {
     match from {
         CLRepr::Array(inner) => Value::from_seq_object(JinjaSequenceObject { inner }),
@@ -122,9 +217,12 @@ pub fn to_minijinja_value(from: CLRepr) -> Value {
         CLRepr::Int(v) => Value::from(v),
         CLRepr::Bool(v) => Value::from(v),
         CLRepr::Null => Value::from(()),
-        CLRepr::PyFunction(_) => panic!(
+        CLRepr::PyExternalFunction(inner) | CLRepr::PyFunction(inner) => {
+            Value::from_object(JinjaPythonFunction { inner })
+        }
+        CLRepr::JsFunction(_) => panic!(
             "Converting from {:?} to minijinja::Value is not supported",
-            CLReprKind::PyFunction
+            CLReprKind::JsFunction
         ),
     }
 }